Added a simple description of the System key
After extensively looking through the docs, it was not clear of the purpose of the System key.  I was looking for the Mac Command key, and found nothing.
I only discovered how to utilize the Command key after searching through the source code.  So I thought it would be beneficial to help others by adding this information!